+++ src/library/Hasql/Mapping.hs
@@ -4,8 +4,12 @@
 module Hasql.Mapping
   ( IsScalar,
     IsStatement,
+    IsTransaction,
+    IsSession,
   )
 where
 
 import Hasql.Mapping.IsScalar (IsScalar)
+import Hasql.Mapping.IsSession (IsSession)
 import Hasql.Mapping.IsStatement (IsStatement)
+import Hasql.Mapping.IsTransaction (IsTransaction)